refactor(chaintx): Lift the chainable fixtures out of `mod tests`
A benchmark target is an external crate, so nothing under `#[cfg(test)]` is
reachable from one. Move `predecessor_resolving`, the chainable-pair
constructors and `forge` into a `test-helpers`-gated `chaintx::test_helpers`,
so the tests that own these fixtures and the benchmark that measures them
build the same thing rather than two fixtures that have to be kept in step.

Adds `fan_in_link_primitive_witnesses`: one successor funded entirely by `n`
predecessors, one thruput apiece -- the shape an N-way chain collapses from.
Also adds `ChainWitness::link_kernel`, mirroring `CastWitness::link_kernel`,
since a caller that proves a chain needs the kernel to pair with the proof and
`new_kernel` is `pub(super)`.
