+++ neptune-consensus/src/chaintx/update.rs
@@ -691,7 +691,7 @@ pub(crate) mod tests {
     use tasm_lib::twenty_first::util_types::mmr::mmr_trait::Mmr;
 
     use super::*;
-    pub(super) use crate::chaintx::chain::tests::forge;
+    pub(super) use crate::chaintx::test_helpers::forge;
     use crate::chaintx::link_primitive_witness::LinkPrimitiveWitness;
     use crate::chaintx::mock_single_proof_digest;
     use crate::proof_abstractions::tasm::builtins as tasm;
