+++ neptune-consensus/src/transaction/validity/tasm/single_proof/fix_branch.rs
@@ -276,8 +276,8 @@ pub(crate) mod tests {
 
     use super::*;
     use crate::chaintx::cast::CastWitness;
-    use crate::chaintx::chain::tests::deterministic_chainable_link_primitive_witnesses;
-    use crate::chaintx::chain::tests::forge;
+    use crate::chaintx::test_helpers::deterministic_chainable_link_primitive_witnesses;
+    use crate::chaintx::test_helpers::forge;
     use crate::chaintx::mock_single_proof_digest;
     use crate::consensus_rule_set::ConsensusRuleSet;
     use crate::proof_abstractions::tasm::builtins as tasm;
